requirements.txt
dnspython
flask
flask-httpauth
ujson
loguru
